standalone/requirements.txt
telethon
python-binance==1.0.14
pyyaml
pysimplegui
pandas
pyinstaller